core/requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.4
tzdata==2025.2